Fixed a build error that could occur when attempting to build without a .git directory present. To workaround, this required renaming the 'config' directory to 'codequality' instead.
The root cause of this, is a bug in JGit, which is used by the Nebula plugin. The problem occurs when there is no '.git' directory, such as when using 'Download Zip' from GitHub instead of 'git clone'. The gradle plugin attempts to find a configuration file, and incorrectly attempts to load the 'config' as a file.
This bug is specific to the name 'config'. Renaming the directory works around the problem, allowing builds to succeed when building source from a zip archive.