WIP updated analysis algorithm
analyzePackage now collects all MXF track files (and essence partitions), as well as all CPLs in maps, before actually validating. The sequence namespace for each track file is retrieved from the list of CPLs, and passed on for track file type-specific validation. After that, the CPLs are validated - with the header partitions being passed along for essence descriptor matching.